Set to style of package
