+++ README.md
@@ -1,4 +1,4 @@
-## TkTerminal
+## `tkterm`
 ![image](https://github.com/littlewhitecloud/TkTerminal/assets/71159641/b5354357-a28c-4cbe-b13c-2c6f06caeb19)
 ![image](https://github.com/littlewhitecloud/TkTerminal/assets/71159641/9fb5a8bc-0782-4ceb-871f-75a8ae18c6d4)
