+++ pyproject.toml
@@ -1,0 +1,10 @@
+[tool.black]
+line-length = 88
+
+[tool.isort]
+line_length = 88
+profile = "black"
+multi_line_output = 3 
+
+[tool.ruff]
+line-length = 125
