+++ tkterm/__init__.py
@@ -1,0 +1,1 @@
+from tkterm import Terminal  # noqa: F401
