+++ upload_requirements.txt
@@ -1,0 +1,4 @@
+setuptools
+twine
+wheel
+build
